Changed to use router
