+++ cdk/react-chatbot/package.json
@@ -35,6 +35,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-markdown": "^9.0.1",
+    "react-router-dom": "^7.0.1",
     "react-scripts": "^5.0.1",
     "react-simple-code-editor": "^0.14.1",
     "react-syntax-highlighter": "^15.6.1",
