+++ cdk/react-chatbot/src/index.js
@@ -1,15 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
 import reportWebVitals from './reportWebVitals';
+import App from './App';
+
+import { BrowserRouter } from 'react-router-dom';
 import { Authenticator } from '@aws-amplify/ui-react';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <Authenticator.Provider>
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
     </Authenticator.Provider>
   </React.StrictMode>
 );
