+++ cdk/react-chatbot/src/pages/Logout.js
@@ -1,0 +1,16 @@
+// a logout react component
+import React from 'react';
+
+const LogoutPage = () => {
+
+    return (
+        <>
+            <h1>Logout</h1>
+            <p>
+                You are now logged out. <a href='/login'>Back to login.</a>
+            </p>
+        </>
+    );
+}
+
+export default LogoutPage;
